cli: only check license file in enterprise version
