+++ cli/internal/cmd/BUILD.bazel
@@ -25,6 +25,7 @@ go_library(
         "iamdestroy.go",
         "iamupgradeapply.go",
         "init.go",
+        "license_oss.go",
         "log.go",
         "maapatch.go",
         "mini.go",
@@ -78,7 +79,6 @@ go_library(
         "//internal/helm",
         "//internal/kms/uri",
         "//internal/kubecmd",
-        "//internal/license",
         "//internal/logger",
         "//internal/maa",
         "//internal/retry",
